+++ package.json
@@ -11,7 +11,6 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "ali-oss": "^6.0.1",
     "axios": "^0.18.0",
     "babel-polyfill": "^6.26.0",
     "debug": "^2.6.3",
@@ -28,7 +27,6 @@
     "koa2-cors": "^2.0.6",
     "mysql": "^2.15.0",
     "nunjucks": "^3.1.3",
-    "socket.io": "^2.0.4",
     "svg-captcha": "^1.3.11",
     "swagger-injector": "^4.0.0"
   },
