+++ routes/api/test.js
@@ -1,30 +1,25 @@
 const router = require('koa-router')();
-const {createToken, checkToken} = require('../tool/token');
+const axios = require('axios');
 
-router.prefix('/yun/test');
+router.prefix('/blog/test');
 
-router.get('/get', async (ctx, next) => {
-    let token = await createToken({id: 1});
-    console.log(token.data);
-    ctx.DATA.data = {
-        ...token.data.data
-    };
-    ctx.DATA.message = 'this is a get test and jenkins!';
-    ctx.body = ctx.DATA;
+/**
+ * GET TEST
+ */
+router.get('/get', async function (ctx, next) {
+  let dat = await axios.get('http://127.0.0.1:3000/core/test/get');
+  ctx.DATA.data = dat.data;
+  ctx.DATA.message = 'This is the GET test.';
+  ctx.body = ctx.DATA;
 });
 
-router.get('/token', checkToken, async (ctx, next) => {
-    console.log(ctx.res.USER);
-    ctx.DATA.data = ctx.res.USER;
-    ctx.DATA.message = 'this is a get test and jenkins!';
-    ctx.body = ctx.DATA;
-});
-
+/**
+ * POST TEST
+ */
 router.post('/post', function (ctx, next) {
-    ctx.DATA.data = ctx.request.body;
-    ctx.DATA.message = 'this is a post test and jenkins!';
-    ctx.cookie.set('MMS30', 'test')
-    ctx.body = ctx.DATA;
+  ctx.DATA.data = ctx.request.body;
+  ctx.DATA.message = 'This is the POST test.';
+  ctx.body = ctx.DATA;
 });
 
 module.exports = router;
