+++ routes/database/index.js
@@ -9,16 +9,16 @@ const mysql = require('./connection');
  * @returns {Promise} 该表所有列
  */
 const findAll = function (tablename) {
-    return new Promise(function (resolve, reject) {
-        var sql = 'select * from ??';
-        mysql.execQuery(sql, [tablename], function (err, rows) {
-            if (err) {
-                reject(err);
-            } else {
-                resolve(rows);
-            }
-        })
-    });
+  return new Promise(function (resolve, reject) {
+    var sql = 'select * from ??';
+    mysql.execQuery(sql, [tablename], function (err, rows) {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(rows);
+      }
+    })
+  });
 };
 
 /**
@@ -28,16 +28,16 @@ const findAll = function (tablename) {
  * @returns {Promise} 返回
  */
 const find = function (tablename, require) {
-    return new Promise(function (resolve, reject) {
-        var sql = 'select * from ?? where ?';
-        mysql.execQuery(sql, [tablename, require], function (err, rows) {
-            if (err) {
-                reject(err);
-            } else {
-                resolve(rows);
-            }
-        })
-    });
+  return new Promise(function (resolve, reject) {
+    var sql = 'select * from ?? where ?';
+    mysql.execQuery(sql, [tablename, require], function (err, rows) {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(rows);
+      }
+    })
+  });
 };
 
 /**
@@ -48,16 +48,16 @@ const find = function (tablename, require) {
  * @returns {Promise} 修改状态
  */
 const update = function (tablename, one, twe) {
-    return new Promise(function (resolve, reject) {
-        var sql = 'update ?? set ? where ?';
-        mysql.execQuery(sql, [tablename, one, twe], function (err, rows) {
-            if (err) {
-                reject(err);
-            } else {
-                resolve(rows);
-            }
-        })
-    });
+  return new Promise(function (resolve, reject) {
+    var sql = 'update ?? set ? where ?';
+    mysql.execQuery(sql, [tablename, one, twe], function (err, rows) {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(rows);
+      }
+    })
+  });
 };
 
 /**
@@ -66,15 +66,15 @@ const update = function (tablename, one, twe) {
  * @returns {Promise}
  */
 const op = function (sql) {
-    return new Promise(function (resolve, reject) {
-        mysql.execQuery(sql, [], function (err, rows) {
-            if (err) {
-                reject(err);
-            } else {
-                resolve(rows);
-            }
-        })
-    });
+  return new Promise(function (resolve, reject) {
+    mysql.execQuery(sql, [], function (err, rows) {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(rows);
+      }
+    })
+  });
 };
 
 /**
@@ -87,39 +87,39 @@ const op = function (sql) {
  * @returns {Promise}
  */
 const findLimit = (tabName, field = ['*'], page, num, order) => {
-    return new Promise(function (resolve, reject) {
-        let count_sql = 'select count(id) as total from ' + tabName + ' where is_del = 0';
-        mysql.execQuery(count_sql, [], function (err, rows) {
-            if (err) {
-                reject(err);
-            } else {
-                // resolve(rows);
-                let total = rows[0].total;
-                let red = (page * num) - num;
-                let mon_sql = `select ${field.toString()} from ${tabName} where is_del = 0 ${order && 'order by id desc'} limit ${red}, ${num}`;
-                mysql.execQuery(mon_sql, [], function (err, rows) {
-                    if (err) {
-                        reject(err);
-                    } else {
-                        let dat = {
-                            pageIndex: page,
-                            pageSize: num,
-                            pageCount: Math.ceil(total/num),
-                            total: total,
-                            list: rows
-                        };
-                        resolve(dat);
-                    }
-                });
-            }
+  return new Promise(function (resolve, reject) {
+    let count_sql = 'select count(id) as total from ' + tabName + ' where is_del = 0';
+    mysql.execQuery(count_sql, [], function (err, rows) {
+      if (err) {
+        reject(err);
+      } else {
+        // resolve(rows);
+        let total = rows[0].total;
+        let red = (page * num) - num;
+        let mon_sql = `select ${field.toString()} from ${tabName} where is_del = 0 ${order && 'order by id desc'} limit ${red}, ${num}`;
+        mysql.execQuery(mon_sql, [], function (err, rows) {
+          if (err) {
+            reject(err);
+          } else {
+            let dat = {
+              pageIndex: page,
+              pageSize: num,
+              pageCount: Math.ceil(total / num),
+              total: total,
+              list: rows
+            };
+            resolve(dat);
+          }
         });
-    })
+      }
+    });
+  })
 };
 
 module.exports = {
-    findAll: findAll,
-    find: find,
-    update: update,
-    op: op,
-    findLimit: findLimit,
+  findAll: findAll,
+  find: find,
+  update: update,
+  op: op,
+  findLimit: findLimit,
 };
