+++ routes/tool/error/errorMsg.js
@@ -1,23 +1,23 @@
 module.exports = Object.freeze({
-    // en
-    // 0: 'fail',
-    // 1: 'validation error',
-    // 200: 'ok',
-    // 400: 'invalid param',
-    // 401: 'unauthorized',
-    // 403: 'forbidden',
-    // 404: 'not found',
-    // 500: 'internal server error',
-    // 503: 'service busy',
+  // en
+  // 0: 'fail',
+  // 1: 'validation error',
+  // 200: 'ok',
+  // 400: 'invalid param',
+  // 401: 'unauthorized',
+  // 403: 'forbidden',
+  // 404: 'not found',
+  // 500: 'internal server error',
+  // 503: 'service busy',
 
-    // zh-cn
-    0: '失败',
-    1: '验证码错误',
-    200: '成功',
-    400: '请求出错',
-    401: '未授权的请求',
-    403: '禁止：禁止执行访问',
-    404: '找不到：请检查URL以确保路径正确',
-    500: '服务器的内部错误',
-    503: '服务不可用'
+  // zh-cn
+  0: '失败',
+  1: '验证码错误',
+  200: '成功',
+  400: '请求出错',
+  401: '未授权的请求',
+  403: '禁止：禁止执行访问',
+  404: '找不到：请检查URL以确保路径正确',
+  500: '服务器的内部错误',
+  503: '服务不可用'
 });
