--- routes/tool/oss.js
@@ -1,6 +1,0 @@
-const OSS = require('ali-oss');
-const conf = require('../../config');
-
-module.exports = function() {
-    return new OSS(conf.oss);
-};
